+++ exporter/dynatraceexporter/serialization/serialization.go
@@ -153,8 +153,8 @@ func escapeDimension(dim string) string {
 func NormalizeString(str string, max int) (normalizedString string, err error) {
 	normalizedString = reNameDisallowedCharList.ReplaceAllString(str, "_")
 
-	// Strip Digits and underscores if they are at the beginning of the string
-	normalizedString = strings.TrimLeft(normalizedString, "._0123456789")
+	// Strip Digits if they are at the beginning of the string
+	normalizedString = strings.TrimLeft(normalizedString, ".0123456789")
 
 	if len(normalizedString) > max {
 		normalizedString = normalizedString[:max]