+++ receiver/stanzareceiver/converter.go
@@ -1,4 +1,4 @@
-// Copyright 2019, OpenTelemetry Authors
+// Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,11 +17,12 @@ package stanzareceiver
 import (
 	"fmt"
 
-	"github.com/observiq/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-log-collection/entry"
 	"go.opentelemetry.io/collector/consumer/pdata"
 )
 
-func convert(obsLog *entry.Entry) pdata.Logs {
+// Convert a stanza-style entry to a pdata.Logs
+func Convert(obsLog *entry.Entry) pdata.Logs {
 	out := pdata.NewLogs()
 	logs := out.ResourceLogs()
 	logs.Resize(1)
@@ -37,10 +38,6 @@ func convert(obsLog *entry.Entry) pdata.Logs {
 
 	rls.InstrumentationLibraryLogs().Resize(1)
 	ills := rls.InstrumentationLibraryLogs().At(0)
-
-	il := ills.InstrumentationLibrary()
-	il.SetName(typeStr)
-	il.SetVersion(verStr)
 
 	lr := pdata.NewLogRecord()
 	lr.SetTimestamp(pdata.TimestampUnixNano(obsLog.Timestamp.UnixNano()))